pyproject.toml
[project]
name = "ft-rag-against-the-machine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.1",
    "flake8>=7.3.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=0.1.0",
    "transformers>=4.51.0",
    "mypy>=1.19.1",
    "rank-bm25>=0.2.2",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "unstructured>=0.21.5",
    "bm25s>=0.3.2.post1",
    "pystemmer>=3.0.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

